refactor(ai): near-gate splits for colonial/diplomacy/economy (Refs #3972)
Topic-split colonial naval lite/civilian, diplomacy pass helpers, and
orchestrator economy build so near-gate planners sit under ~750 lines.
